requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-phonenumber-field==5.2.0
kombu==5.1.0
phonenumbers==8.12.31
Pillow==8.3.2
prompt-toolkit==3.0.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
